Class News
A dirty hack was added to app.js that defines a noContent() method. I
defined it there to use in all controllers in the API returns back a 204
no content. However, its not working correctly. When trying to press the
back button from the noContent page, it just reloads the current page
and the second button press takes you back to the class homepage. I fear
it has something to do with the noContent not rending a handlebars
templates, but I'm not sure.